Completed PhoneNumber by using the Result monad.
Added a result monad, and used it to unwrap a series of functions that validate a phone number.

This is much cleaner and more readable than the previous implementation, and also removes duplicatino I had for checking the exchange code.

Also updated to the latest, they had changed the tests to look for "letters" instead of "alphanumerics".